Diagnostic plots for gMAP analyses

Diagnostic plots for gMAP analyses

Usage

# S3 method for class 'gMAP'
plot(x, size = NULL, ...)

Arguments

x

gMAP object

size

Controls line sizes of traceplots and forest plot.

...

Ignored.

Value

The function returns a list of ggplot objects.

Details

Creates MCMC diagnostics and a forest plot (including model estimates) for a gMAP analysis. For a customized forest plot, please use the dedicated function forest_plot.

Customizing ggplot2 plots

The returned plot is a ggplot2 object. Please refer to the "Customizing Plots" vignette which is part of RBesT documentation for an introduction. For simple modifications (change labels, add reference lines, ...) consider the commands found in bayesplot-helpers. For more advanced customizations please use the ggplot2 package directly. A description of the most common tasks can be found in the R Cookbook and a full reference of available commands can be found at the ggplot2 documentation site.